implement custom exceptions

+++ src/main/java/org/example/bankingsystem/services/UserService.java
@@ -7,12 +7,7 @@ import org.example.bankingsystem.repository.UserRepository;
 import org.example.bankingsystem.requests.LoginRequest;
 import org.example.bankingsystem.requests.UserRequest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
 import java.util.Optional;
 
 @Service
@@ -58,12 +53,15 @@ public class UserService {
             oldUser.setPassword(request.getPassword());
             return userRepository.save(oldUser);
         }
-
-        return null;
+        throw new UserNotFoundException("user not found with id : "+id);
     }
 
     //method to delete user
     public void deleteUser(Integer id){
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(
+                    "User not found with id : " + id);
+        }
         userRepository.deleteById(id);
     }
 }
